Fixed Stack overflow in Android model copy.
Android plugin 2.5 broke the proxy-based cache we have
in Android Studio by making OutputFile recursive. The
method 'getMainOutputFile' in the implementation of
OutputFile returns 'this', making the creation of the
cache throw an StackOverflowError. This exception is
thrown when Gradle sync finishes and the IDE attempts
to serialize the Gradle models to disk.

The fix for this bug involves
1. Move away from proxy-based cache. Fixing this
   caching mechanism would be a waste of effort, since
   we have already decided to replace the proxies with
   real classes. Using the proxy-based cache is significantly
   slow, because to access each method from the proxy
   object we need to invoke Method#toGenericString which
   is computed every time and cannot be cached.
2.  Make the new class-based cache be aware of possible
   recursions and use an internal class to detect duplicated
   objects.

This CL fixes #2 (which needs #1 in place.) Now all the
model implementations have a super class that takes the
internal model cache where the models are stored right
at the moment of creation, and never creates a new
implementation directly, but instead, checks in the
internal cache first to see if an similar instance was
already created.

Bug: 37324890

Test: JUnit test.
Change-Id: Ia3b5db1cd7231ff836f6427d2c5e292e66f8ad81