Adding RPC diff validation to perfd-host test.
This change adds a client and service interceptor to the gRPC calls of
perfd-host. The interceptors cache the request and response of each call
as it is being made. When the test is done running a validation is run
on the output of the calls to verify that the RPC calls being made match
the expected RPC calls. If a new test is added, or RPC calls are changed
a new expected output can be generated by adding the VM option
"-Dtest.export.grpc=True"

Test: Perfd-Host
Change-Id: I2e3f718308843b021fee26d4d050ba09f3008a64