Add RESTART deploy type
Problem: App is deployed at state A.
User makes a change that can be warm swapped and runs apply changes.
user then reverts the app back to state A and hits apply changes again.

Gradle thinks the app has no changed so returns no artifacts.
InstantRunBuildAnalyzer incorrectly returns NO_CHANGES.

Note this situtation only occured when editing static methods due ot a quirk:
adding or changing a static method is a COLDSWAP but removing is a HOTSWAP.
That's been modified to always be a COLDSWAP.

Solution:
For full discussion, see bug. In summary we want to restart the app to
remove the patched changes and get the app back to state A. It doesn't
cleanly map to any of the existing deploy type so adding a RESTART
deploy type that just restarts the application on device.

So if gradle tells us a change is not hotswappable but gives us no artifacts,
then we restart the application.

Fixes: https://b/37150413
Test: added test case for restart in InstantRunNotificationTest, created
InstantRunBuildAnalyzerTest
Change-Id: I8844414256c72e4866ce0e62a8dfbb0c9b50c8b3