[APK Debugging] Added mechanism to find and refresh nodes in the 'Android' view.
This CL adds the class AndroidViewNodes. This class makes it
possible to search for a node in the 'Android' view's tree (given
some search criteria) and refresh the contents of the matching
node. This functionality is needed when setting the debuggable
.so file for the APK, through the library editor. If the debug
symbols in the .so contain source folders, those source folders
need to appear in the 'Android' view.

This CL also removes code duplication by abstracting common code
in the class LibraryFolder.

Test: JUnit tests.
Change-Id: I3f98ded14498fd1e134c881aa242a2391c38794e