Remove unused code and update to use feature modules
This CL contains multiple interrelated updates:

* Remove code generating now unused template parameters
(mainly to do with the generation of splits and injecting
parameters in to the base split / lib's manifest).
* Update code that looked in the base split for e.g.
launch URLs to now search all feature splits for the same.
* Change the new module wizard to no longer attempt to set
up dependencies as this is no longer easy to do in an
unambiguous way.
* Change Instant App Projects to use the latest build
tools including previews

Test: existing tests cover code - updated in CL to follow
Change-Id: Ia61431ad99467907bb5d0ebc3fedfd16bac56ce8