Fix starvation in CPU profiler that causes Studio to freeze
Running the app again while we are profiling it causes ClientData
to change its MethodProfilingStatus to OFF without removing
the corresponding LegacyProfilerRecord from the map.

In this scenario, we were never calling myStopLatch#countDown and
checkAppProfilingState was waiting forever for myLegacyProfilingLock
to be released.

We should check for MethodProfilingStatus.OFF, not only if the map
contains the record, to make sure the app is being profiled or not.

Bug: 37302749
Test: Manual testing. After the fix, I can't reproduce the bug anymore.
      There are no automated tests for StudioLegacyCpuTraceProfiler yet.

Change-Id: I9006dfc210dd75f53febc32092c0d33561e108cb