Reintroduce the "use new sync" UI option, under a flag.
This CL reverts commit 31d7ac0157707a99b95724bf7a943a2f610613b0,
and reintroduces the "use new sync" UI option in the
Preferences/Settings dialog, with the difference that
the IDE needs to be started with the system property
'show.new.sync.preference'. Only when this system
property is present, the UI will show the checkbox. This
useful for testing, when we need to switch between
'old' and 'new' sync implementations.

Test: N/A.

Change-Id: Ibda6d8523f25ec8cabd5c357b2aa7a038abcf17c