Group inspector properties table
Current properties table is a dump of all properties we get from the
framework on a node. It's disorganized and hard to read.

Improve it by grouping them into categories based on the property name
and reusing PTable element to display properties into expandable
categories. Inspector properties panel now has a similar same look and feel
as Layout Editor except we have no use for the "Star" feature or the
ability to edit the properties.

Those will be added in the future but requires UX (for Starring) or
workign with framework team to figure out live editing.

Fixes: http://b/37140023
Test: Updated PropertiesTablePanelTest
Change-Id: If698a1fefe39cd66789bf1df64c0c69a3caa623b