New Project Wizard: Fix for NewModuleTest@createNewAppModuleWithDefaults
When adding a new module, in the step that adds a new activity, the old
wizard shows a tile of "Add an Activity to Mobile" while the new wizard
was showing "Add an Activity to Tablet and Mobile".
Before running the test, make sure to change
WizardUtils@isNpwModelWizardEnabled to always return true.

Test: NewModuleTest@createNewAppModuleWithDefaults
Change-Id: I79b669cb1260a437f4737b42cc36b9d0ba1b30df