New "Add Sample Data" dialog
This dialog offers a few data sources to add sample data from. It's just
a quick way to start adding some data.
Currently the dialog is only accesible from the "Add Sample Data" action
but the next step is adding an inspection that provides a suggestion for
adding sample data when certain conditions are met.

Test: Pending
Change-Id: Iaa39b32c3c6cb8be1ab1d2db2789f83a3b1bed5e