[AIA] Consider features as libraries for tests
We still don't support running instrumented tests against instant apps
in a realist way, simulating the entire instant apps environment. For
the moment, tests should be run for each library as it's done for
android library modules.

This CL removes the error in the run configuration for feature (and
base) modules by adding it to the supported types of projects.

Test: Unit test

Bug: b/37562629
Change-Id: I217d94a4fa8e45d2ad2a8e8fdfcba21176962efb