Update lint's findViewById detector to handle O SDK implicit casts
(IDE side corresponding to lint changes)

The findViewById signature was updated in Java 8 to implicitly
cast to the type of the LHS. This makes code easier to write, but
it means that the lint check (which was checking the explicit cast)
was no longer enforcing consistency between the type and the tags
used in the corresponding layout.

This CL updates lint to handle this scenario correctly; it now
checks the type of the surrounding expression and uses that inferred
type to perform the compatibility check. It also cleans up the code
a bit and switches from a DOM parser to a pull parser, etc.

It also adds a new check to warn about scenarios where the cast is
incompatible in 1.8 (and adds a quickfix to add the missing cast.)

This CL also adds some lint infrastructure improvements needed for
this fix: first, quickfix descriptors can now request format and
symbol shortening, and the gradle model mocking can request language
levels.

Change-Id: I33bf88c1563261af4b5183e2111af15c669df9b6
Tests: Unit tests included