Profile at most a process at a time.
A regression from ag/2008848 let profilers monitor multiple
processes but this feature is not ready for our users to use.

With this CL, we stop profiling a process when any of the
following happens.

(1) setting a new process
(2) setting a new device
(3) the process being profiled became dead

This logic may change depending on our decision of how to
let users to stop profilers (b/37193922).

Bug: b/37243164

Test: manual; verified by logging in Studio and perfd.
An automatic test will be added soon.

Change-Id: If963a1a50efc8c20e5f9ac1d1f10d670c0c8c16f