Improve Template Test coverage
Template Test had several restrictions in place on its
coverage of templates in order to reduce its running time.
Since we have removed unnecesarry syncing the running
time is now well under the acceptable maximum of ten
minutes.

This CL re-instates a much fuller iteration of the
available parameters and hence exercises a much larger
part of the template system at the cost of an increased
run-time.

Test: None
Change-Id: Ia1134ee4b4459e795b5a135423a213320a6373de