Add ability to render preview for sub trees.
This feature is only enabled when the device is still connected because
we need to recapture a new preview image from the app.

User can double click on a preview component or right click on a node in
the layout tree to show only the selected sub tree. This allows devs to
isolate a portion of their UI for inspection.

Fixes: b/38201477
Test: added LayoutTreePanelTest
Change-Id: Id406dda447d489ff3b8d79bd37b075e5da898a11