Add simpleperf as a profiling option in CPU profiler
This CL introduces the concept of profiling preferences, that contains
the profiler type (ART or simpleperf), profiling mode (sampled or
instrumented), buffer size, sampling interval and a name to identify
the preference.

Two default preferences are created to represent ART's sampled and
instrumented profiling, so we still maitain these types of profiling.
Another default preference was added under a flag to represent
simpleperf profiling (O+ devices only).

The design is extensible and later the users will be able to add custom
preferences as well.

Bug: 37946452
Bug: 37294023
Test: Added unit tests

Change-Id: I126126685758eebf979a0931cc84e4af68b7b260