Rename updateContentsInFile to searchAndReplace to match upstream.
Upstream (tools/base/.../PerformanceTestProjects.java) uses the
name `searchAndReplace` for this functionality.  Using a different
name in our source makes the function less discoverable (when looking
for the equivalent) and also increases cognitive load when doing a
mental diff.  For this reason, this diff changes the name to match
the upstream variant.

Test: Junit GradleSyncPerfTest.testGradleSyncPerf()
Change-Id: If33343869a17bff33891414c23eb82cf44afa4d8