Use ITEM_MAP_LOCK in MultiResourceRepository
This is an attempt to fix NPEs that started to show up in our crash
reports. Our current locking scheme is inconsistent and broken and will
have to rewritten from the ground up, but I hope this change will stop
the bleeding.

Using both "this" and ITEM_MAP_LOCK (which is static and shared be all
repositories) can easily result in deadlocks. That's because methods in
AbstractResourceRepository grab ITEM_MAP_LOCK before calling
MultiResourceRepository methods. On the other hand
MultiResourceRepository calls methods on children (while holding the
"children lock") which in turn need ITEM_MAP_LOCK.

Test: existing tests, as we don't have a repro scenario.
Bug: 37967214
Change-Id: I9bbb7b308a1c63f927c318a5923a2ce06dff908d