[AIA] Fixes deploy task to read devman output
To know if deploy has worked or not, Studio has to read output from
Devman, that is shown in the Logcat. The task should succeed or fail
based on this output.

This CL reads the output and notifies the user any errors.

TODO: migrate deploy to a separate package (b/38154760)

Test: current tests should pass

Bugs: b/37708612 and Studio part of b/36966779.
Change-Id: I3b99e66524c81d514f43991dd1cabe988c3842ba