Initial version of the CPU profiling configurations dialog
Instead of having only "Sampled" and "Instrumented" profiling options,
users should be able to create and set their own profiling configurations.
They can define the sampling interval, profiler type (ART or simpleperf),
and (for pre-O devices only) the maximum size of the buffer file that
contains the recording output.

Users configurations are saved on per project basis and should be
persistent, so users don't need to recreate them every time they open
Android Studio. This CL adds a dialog that can be used to edit such
configurations.

Simpleperf configurations were also added in this CL under a flag. If
the flag is disabled, the user can't create, edit or view simpleperf
configurations. If it is enabled, users can edit, create and view
simpleperf configurations, but they are only listed (and can be used)
for O+ devices.

Test: CpuProfilerStageTest profiling modes tests
Bug: 29258733
Bug: 37294023
Change-Id: Ie8b22d40238513ca92e60b775b7febb55efa45a6