IDE no longer applies AndroidFacet to root module on sync failures.
This behavior was introduced to avoid having the "Android
framework found" message on Gradle-based projects. This is
no longer needed. The most important reason for the removal
is that our editors work with AndroidFacets, under the
correct assumption that AndroidFacet identifies a module
as an "Android module" and having
an AndroidFacet on a module that is not Android (the
top-level module) is plain wrong.

Test: JUnit test.
Change-Id: I3afee41a95dd229a4a3cdcb6cacb7bcb6da4fd30