Hide live tracking behind the profiler.livealloc flag
This CL puts the studio-side live allocaiton tracking functionality
behind an additional flag (on top of profiler.jvmti). With the two
flags enabled, allocation tracking now starts immediately on entering
memory stage, and finishes on exiting the stage. Also the recording
button is hidden so the tracking is always enabled if the user stays
inside memory

Bug: 38312813
Test: Unit test

Change-Id: I70a513a36c3594d3fe5908800d487fe41e3c3bc2