Do a build+warmup run prior to running GradleSyncPerfTest
Previously, we were experiencing issues where the performance graph
did not align with observations in AndroidStudio.  This change
attempts to mitigate the discrepancies by building the project
and warming up the JVM prior to recording the perf metrics for
display on the performance dashboard.

Long term, we still want to collect more performance metrics and
run in a full GUI version of AndroidStudio, but this CL is probably
a good stopgap measure.

Test: Ran the performance test, used the AnalyticsInspector to
verify that exactly one gradle sync was in the resulting trk file.

Change-Id: I924091451d5feb7228d955e394286c52da878dff