Ensure device is ready before launching perfd
Previously studio profiler pushes and runs perfd on device
as soon as it becomes ONLINE (as reported by ddmlib). Turns
out that can be too soon in userdebug builds, and occasionally
the file system can still be in read-only mode. Added a check
in the perfd launching thread to wait till the device is fully
ready by checking for the dev.bootcomplete flag on device.

Bug: 38400565
Test: Integration test to come

Change-Id: I82a3bc094ed609c7714300d6162c31a15416522b