Temporarily ensuring non-null model in DesignSurface
Right now in one case we'll set the model of the preview panel's design
surface to null temporarily, just so the progress spinner will show.
This no longer works, since we rely on the surface's model for component
creation.
Disabling the setting of the model to null temporarily until I can
figure out how to make the model truely non-null and still get the
spinner to show.

Test: fixes OpenIncludedLayoutTest, EditorFixtureTest
Change-Id: I604bff6df322e8cb9517d5cd8edd70e3c53a4d93