Workaround problem in gradle dependency caching
The ModuleClassLoader would get the library dependencies by using
AndroidModule.getExternalDependencies while the resource repositories
were using AndroidModuleModel.getSelectedMainCompileDependencies.
Because of a caching issue, it seems that getExternalDependencies was
returning stale content and this caused a mismatch of the libraries used
that was causing layoutlib to discard them.
This CL unifies the dependency discovery to workaround the problem.

Bug: 62024809
Test: Existing tests should pass
Change-Id: Icad1f435c26f95306ad4d02923c43e1bdc910751