[AVD Mgr] Search for device definitions by API order
Virtual device definitions are in 'devices.xml' under
system-images/android-nn/, with potentially conflicting versions in
the various API-level subdirectories. We should give precedence to the
definitions in the higher-valued API subdirectories, because the files
there will have the latest updates.

In DeviceManager.initSysImageDevices(), look at files in the desired
order.

Bug: 37323877

Test: Modified DeviceManagerTests.java

Change-Id: I34e79040045df9ce1427987a39981c35219676ed