Fix Kotlin androidTest are not recognized as androidTest
The issue is that a KtClass is not an instance of PsiClass, so
the android test configuration did not think it should activate
itself.

The fix is to fallback to the UAST when looking for PsiClass
(or PsiMethod) fails.

This change fixes the part about running tests in a simple
application, but there is more code in the "instrumentation" package
that probably needs to be fixed to work with the Kotlin AST (i.e.
fallback to UAST).

Bug: 38452937
Test: Existing tests?
Change-Id: Ic9c78dfa830929faecbd6a5c1470be94d315b04c