Fix APK editor preview pane after selecting Kotlin file
Null out editor pointer immediately after disposal, to prevent retaining
a pointer to a disposed editor if an exception is thrown later in the
method. This would cause double disposal errors and prevent any further
use of the analyzer.

Bug: 62103513
Test: This adds graceful recovery from erros that shouldn't occur to
begin with. Will add tests when the root cause of the errors are fixed.
Change-Id: I57f71efe5ba321225379056405e676f703f1684d