Merge legacy and O+ allocation contexts queries
- GetLegacyAlloationContexts and GetAllocationContexts both now
return AllocationContextsResponse.
- Fixed an issue in which class ids can collide between multiple
apps. As we used a tracker/converter per process, but we ignore
the process id when storing data into the database. This change
now include process id + session info for class+stack entries.
- Removed the need to encode stacks as SHA256, which can be expensive.
In O, stack ids are simply int32s representing an index for a
prefix tree.

Bug: 37175866
Test: updated unit tests
Change-Id: I5acac70db16e31a8708716618a5cd35abbca926b