Flush trk file at end of test, require ANDROID_SPOOL_HOME
Previously, there was a bug where the scheduler was not being advanced
at the end of the test, which resulted in an empty TRK file being
generated.  Also, this test is only meaningful/useful if the
ANDROID_SPOOL_HOME is set (defaulting to the project directory,
which gets auto-deleted, is completely useless), so I got rid
of that branch.

Test: Ran the test, verified a non-empty TRK file is generated.
Change-Id: I9d8efceaa2ba8ede3df9740c88ec0518cffe207a