Avoid keeping callstack in memory after parsing a capture
perflib provides an API to get the thread information from a trace file.
VmTraceData#getThreads() return a list of ThreadInfo and we were using
it as the key of our Capture map. The issue is ThreadInfo keeps a
reference to the top-level Call of that thread and, recursively, the
entire callstack is kept in memory.

As we just need thread id and thread name after parsing a capture, we
should use a simple object that wraps these fields as the key of our
captures map.

For a 20MB trace file, we create approximately 1.5 Million nodes. If we
kept the Call objects in memory, we would have an additional of ~100MB,
which can be avoided by using a simple object the key of myCaptureTrees.

Test: existing
Bug: 62248892
Change-Id: I13b99c8d91f55b3cce807737bdf7e327ffb17d7e