Refactor VmTraceParser for CPU profiler models
This CL updates VmTraceParser support so it can parse directly into
custom models. This allows to overcome "out of memory" exception
when a big trace file used in cpu profiler. The reason was that
we were parsing using VmTraceParser into its models then
we converted into our models (i.e from Call model to CaptureNode model).
In a result, at the same time we had a lot of object in memory.

Specifically, for a 86mb trace file: 3million instance of
each "Call" and "Call.Builder" and "CaptureNode", the old version
was crashing with "out of memory" exception.

Now VmTraceParser parses and sends events via VmTraceHandler interface,
and the callers should implement this interface and convert into their
models. Potential benefit of this approach is that we can parse
on the fly without keeping all objects in the memory.

This CL implements the VmTraceHandler for CPU.

Test: http://b/62373380, diffed with the old implementation

Bug: 62369152

Change-Id: I6b2769c2e5bbe0a3ac5148aa5e83438c5cc4638a