Adding a isClosed check to the datapollers.
This change will prevent each poller from ticking if the database,
is in a bad state. An error will get logged the first time the database
attempts to get accessed from a bad state, then each poller will
early out when the connection is closed. The ProfilerService will
detect this bad state and disconnect each poller to prevent spamming
the users log, and crash reporting with errors.

Test: DataStoreTable
Change-Id: I51c03833910e57f474ae4ea439c51c1a1f8e1f5d