Hide "Use same selection" option if IR is enabled.
WHen IR is enabled, we always deploy to the same device anyways. The
checkbox is not respected so displaying it only serves to confuse
the user.

So hide checkbox if the IR is enabled. Note it possible for the
user to have IR enabled but not use IR if their gradle version
or target device api version is too low. But we have no way of
knowning about api version until they select it.

Test: n/a, UI change
Change-Id: I6429a4bfe6220ada0372931ee4953be37369f583