Don't kill running app if instant run is disabled and user hits "Run".
For IR, we kill the last running session on a device when user hits "Run".
Because with IR, we auto deploy to the same device and kill the session
to make IR more reliable.

But we do this even if IR is disabled and we're deploy to a difference device.
This cl limits to killing the session only if IR is enabled.

Change-Id: Ie156d115d7e5156eb6a744415d709133672ae35a
Fixes: b/37137376
Test: n/a