Fix race in GUI test fixture (IdeFrameFixture)
Previously, there was a race condition between waitForBuildToFinish and
the AndroidProjectBuildNotifications subscription firing.  I suspect
there may be a relatively recent (last week or two) regression that
either introduced or exposed this bug.  Waiting an extra ten seconds
for the subscriptions to fire, which should be plenty of time.

Test: Ran GradleSyncGuiPerfTest, which is one of the tests that calls
invokeProjectMake.

Change-Id: I52a51e5ed8a7c25f5741ab54ecfc3325c5192ca7