Fix Network Details View refresh issue
When new HttpData is selected, the contents of the connection
details panel should refresh themselves. However, when new
contents were not as tall as the old contents, Swing didn't know to
clear out the old areas, so we have to tell it to do so explicilty,
will a top-level repaint.

Test: N/A, visual change only
Bug: 37507526
Change-Id: I3966a836a8ca04740530d8b5dc0aa2aa82fbd623