Use embedded Maven repo only when syncing and building newly created projects.
When a project is created using the NPW, the IDE keeps track that
this is a "newly created" project (in GradleProjectInfo.) Sync,
source generation and any subsequent builds will use the embedded
Maven repo. Once the project structure changes and a sync is needed,
the IDE will no longer consider this project as "newly created"
and the project will stop using the embedded repo.

Bug: 38463221

Test: JUnit tests.
Change-Id: I99a17f62711c6b6e28034ec7ded442517b217e59