Separate build from perf test
One of the performance metrics we want to track is the "initial sync" time.
To get accurate real-world numbers, we need to simulate a real-world
configuration to the extent possible.  In particular, when users first
open their IDE, they generally have build artifacts from their previous
uses of AndroidStudio.  We have seen these build artifacts impact sync
time, so a performance test should ideally contain these artifacts when
measuring initial sync performance.

Intuitively, there are two solutions, neither of which are particularly feasible:
1.  Do a gradle build (using the tooling API) prior to launching the performance
    test.  In some respects, this is the cleanest solution.  The problem is that
    it isn't as simple as executing a "gradle assemble" invocation.  There are a
    number of flags set by AndroidStudio to make the build succeed.  Additionally,
    the set of tasks invoked by android studio appears to be very non-trivial to
    calculate (just executing a "gradle assemble", even with the proper flags set,
    will throw errors) - accurately simulating what AndroidStudio would have built,
    and keeping that code in sync with AndroidStudio's actual behavior, imposes an
    unacceptable maintenance burden.
2.  Launch the build (as part of the test) using AndroidStudio.  Android studio
    already has all the code to perform a build, so just invoking this code seems
    like a natural solution.  Unfortunately, there are a few problems/drawbacks.
    Most notably, the code to invoke a build depends on the project structure
    being loaded into the IDE, and loading the project structure requires doing
    a sync.  But the goal is to measure the performance of initial sync with
    artifacts, so we can't perform a sync prior to the "initial sync", as that
    is a contradiction in terms.

The best solution I've been able to identify is to use two UI tests.  The first
one opens+builds the project, and saves the results to a specified location.
The second one launches android studio and opens the specified location.
This allows us to most accurately simulate real user behavior by having real
instances of AndroidStudio doing both the preparation and performance analysis.

Ultimately, the primary goal of the first test is just to generate the data
necessary for the second test, although the test is useful/available as a
stand-alone test.  It would be possible to have the data generation as a non-
test java application, but there aren't really any advantages to such a
configuration and there is the drawback that you don't get to utilize the
existing infrastructure (including reporting of failures on the buildbot page).

One other (incidental) benefit to breaking out the test into separate programs
is that they can be run individually.  You can do the data generation once, and
debug the sync times independently many times using the same generated data.

Test: Run the two tests.
Change-Id: I0eb3303f993b97d0e9bf12d162a22e120da3129f