Replace GradleUtil::getDirectLibraryDependencies with getModuleDependencies.
This is part of the effort to switch to DependencyGraphs in IDE.

Level2Dependencies only returns flattened list of dependencies, so there will be
no way to get direct-only dependencies.

Based on the usages of getDirectLibraryDependencies, what is needed is a merged
list of module dependencies for each variant, because in all use cases, the returned
list is filtered by getProject returns non-null.

Test: existing tests should cover it.
Change-Id: I540993ceaf67f92ed9bdb0631e55080dc53ed594