[Bug27896976] Add a message dialog when attempting to clean while building
It is possible to press the clean project menu while a project build is
running. This causes some issues when both task are running at the same
time. To prevent that, this change first checks that a build is not running
before adding a clean task.

Another option would be to disable the clean project menu when a build
starts and re eanble it once it finishes, but if for some reason the build
does not finishes correctly, this menu item would be left disabled.

Patch 2: Add CleanProjectActionTest.java and use the same ballon as other
notifications instead of a Warning Dialog for consistency.

Test: Add 2 new tests to confir the action calls clean only when build is
not running.

Change-Id: Ib5d77f7a9e2f5cd424e48c8910a6ad53c2a9cb56
Bugs: 27896976