[IAPP] AAPT2 link failed while running Master/Detail Flow template
Strings used on the base manifest need to be on the base fragment. At
the moment only "app_name" was needed, but "Master Detail Flow" needs
one more.
Use "baseFeatureResOut" template variable as the destination of the
extra strings (strings_iapp.xml.ftl). If we are building an instant app,
this variable points to the base resource, if we are building a normal
app, the variable points to the "app" resource location.

Test: Covered by existing template tests and UI tests
Change-Id: I8de728fc227ca0ffb2b83f5b845d6f34b7459b81