Capture should be parsed only once and in a separate thread
Cpu captures were being parsed twice due to a concurrent issue. We were
starting trace capture parsing after stopping the capture. In another
thread, we were starting the parsing of the same trace because
getCapture(traceId) couldn't find the capture corresponding to that
trace (as it was still being parsed).

We need to make sure the capture is parsed only once (no matter if
started by stopCapturing() or getCapture(traceId)). In order to
accomplish that, CpuCaptureParser was created to handle trace parsing,
making sure the traces are parsed only once.

As a result, we are able to parse larger captures because we consume
half of the memory we were using before. Some observed results below:

Before the Fix
* 40MB instrumented capture with 20s duration produced ~3 million
* CaptureNode instances (251MB) plus ~3 million ArrayList instances
* (96MB), resulting in 347MB. Time to parse and display capture was
* ~10.8s.

* 80 MB instrumented capture with 20s duration caused studio to freeze
* for several minutes until eventually crashed due to lack of memory.

After the Fix
* 40MB instrumented capture with 20s duration produced ~1.5 million
  CaptureNode instances (125MB) plus ~1.5 million ArrayList instances
  (48MB), resulting in 173MB (2x improvement). Time to parse and display
  capture was ~1.2s (9x improvement).

* 80MB instrumented capture with 35s duration now uses approximately the
  same memory we were using for 40MB captures before (2x improvement).
  Time to parse and display capture varies. After 5 tries, minimum time
  observed was ~4.4s and maximum ~167s.

* 100MB instrumented capture with 35s duration produces ~3.7 million
  CaptureNode instances (314MB) plus ~3.7 million ArrayList instances
  (120MB), resulting in 434MB. Time to parse and display capture varies.
  After 6 tries, minimum time observed was ~2.9s and maximum ~60s.

Test: Unit tests
Bug: 62527992
Bug: 62527394
Change-Id: I05fec2a6364105c5c2fd7c090da61d2f797a4095