Refactor LegendComponent to share logic between draw / size methods
Previously, LegendComponent had hand-crafted logic in the
getPreferredSize and draw methods which had to be carefully kept in
sync. Tests had to be written to further verify that a code change
wouldn't accidentally break the relationship.

Now, a concept of a "LegendInstruction" is introduced, which contains
information about the pieces of the legend we want to render.
|getPreferredSize| loops through these instructions and builds up
knowledge about the legend's bounds, while |draw| also loops through
these instructions and uses the same logic to request render calls.

Rewrote tests from scratch, since the old tests are obsolete and the new
framework offers some new ways to verify functionality.

This new approach does not adversely affect render time, and in fact
seems to slightly speed things up (although the time ranges are so small
it's not significant either way):

Before change: ~7-15 us
After change:  ~4-12 us

Bug: 36474036
Test: Updated, all non-rendering code tested
Change-Id: Iaf7bff1ff4316c53b6d2d3f5f1e80bf75f4aff7b