[AVD Mgr] Make ConfigureAvdOptionsStepTest shard-able
The unit test ConfigureAvdOptionsStepTest fails when sharded
in Bazel. It succeeds in Bazel when not sharded because
AvdDisplayListTest runs first and provides some required set up.
It succeeds locally when run by itself because the local machine's
SDK is visible.

This CL adds the required set up to this test so it is not dependent
on either a local SDK or on any other tests.

Bug: 62359884

Test: This is a unit test.
To test this test, I commented out AvdDisplayListTest to expose
the problem even when not sharded. Verified that the modified
code passes unit tests locally, in PSQ, and in Bazel.

Change-Id: I99ec1c390fcbd126de5bc38b45f7755812201577