Fixed Gradle sync with AGP 1.5
This CL fixes syncing a project using AGP 1.5 and Studio 3.0.

Since we moved to the IdeModel classes, it is pretty easy to
break Gradle sync with older plugins. As part of this CL, I'm
adding an integration test that syncs a project with
AGP 1.5 and Gradle 2.2. The test is disabled since I need
to include all dependencies in the embedded repo we use to
run tests in the sandbox. The fix for the bug (this CL) needs
to be submitted ASAP.

Bug: 62169504

Test: JUnit test.
Change-Id: I6729476e3f46393746f9e08abc3d704a313fd943