Cleanup CpuTraceDataSeriesTest tests and make it use truth
CpuTraceDataSeriesTest tests intended to use "max range"
(min_double to max_double) in a couple of tests, but they have a typo.
The tests were passing because this range is not relevant for them, but
we need to change it for purity reasons.

We also change these tests to use truth.

Test: updated some tests and they still pass
Change-Id: I01b458c12980b5d58c58c15bf2523e612045890a