Only use styleables from the current namespace.
Before this change we would include attributes from all styleables
that was matching the current view name. For the GridLayout in the
support library it would include the attributes defined for the
framework GridLayout as well as the attributes from the support library.

This approach results completion results that are not valid attributes
for the support library GridLayout.

Bug: 37070953
Test: Updated/added.

Change-Id: Ia028ff19268ddb47d708745d25321551f5cca0ce