"Close" button in Debugger window should allow disconnecting
The Debugger tool window has a "Stop" (Red square icon) button to
force termination of the application being debugging, and a "Close"
(diagonal red cross icon) button.

When clicking the later, the intended behavior is to show a dialog
letting the user choose between "Terminate" (same as "Stop") or
"Disconnect". "Disconnect" means stop the debugging session but leave
the application being debugging running.

* Before this change, the Android Debugger would always terminate the
  application when the "Close' button is clicked (without showing the
  dialog)

* After this change, the Android Debugger implements to intended
  behavior (Dialog + support for "disconnect").

Bug: 37136195
Test: n/a (for now: requires infrastructure work)
Change-Id: Id69c195c396b1197c1f98eb64a0351b3d519b5f5