Update gutter icons when resource changes
Currently, when the user edits an image resource (e.g. changes the
color of a vector image), the gutter icons shown next to lines which
reference the resource do not get updated to reflect these changes.
This issue is caused by the presence of stale entries in
GutterIconCache.

This CL fixes the issue by having GutterIconCache#getIcon check for
changes in image resource files and create new Icons for those
resources if their cache entries are no longer valid. Document
contents are used to generate the new Icons if the resource is open
in an editor so that the new Icon takes into account any unsaved
changes in the resource XML.

GutterIconRenderer#getIcon has been updated to consult GutterIconCache
for all Icon requests, rather than using a potentially stale Icon
stored in a local variable.

Test: Covered by GutterIconCacheTest#testIsIconUpToDate_*
Change-Id: Ibee09d7607945985484f19591a6d44378c39d5f3