Dependency on modules that wraps aar should show up as external dependencies.
Since gradle plugin 3.0, sub-module that wraps aar are returned as
external dependencies, i.e. with type ANDROID_LIBRARY. See ag/2318991.
But the old version of plugins still return such dependencies as module
dependencies.
This change detects such dependency by checking location of aar bundle,
if aar bundle is not inside of build directory of the module, then
it must be an aar wrapper module.

This change also removes artifactFile from IdeLevel2ModuleLibrary to
keep consistent with ModuleLibraryImpl in builder-model.

Add unit test and integration tests with head plugin and 2.2 plugin.
Integration test with 2.2 plugin is passing locally.
But has to be disabled for sandbox. Re-enable when 2.2 is added to prebuilts.

Test: IdeAndroidProjectIntegrationTest, IdeLevel2LibraryFactoryTest.
Change-Id: I4aea04ed2c38ea3948091f12016cdd44f92e1b09