Build and bundle release gradle plugin only if flag set.
Use a new flag to specify when to build and bundle the release
gradle plugin rather than the -dev gradle plugin. Another change
in this topic should make the go/ab build_studio.sh script pass
this new flag, so that the go/ab builds on this branch can be
considered canary candidates with the release gradle plugin.

Bug: 62902841
Test: N/A
Change-Id: Ibe179bdcaba71bdc8647766df57022bd223e0ca6