[Bug 38313922] Show an error message when a dependency cannot be downloaded.
When a gradle sync fails due to a missing dependency, a quickfix to download
such dependency is created. If the dependency is part of ANDROID or GOOGLE
maven repositories, the whole repository is downloaded instead. That causes
that the version of the dependency that is required is never checked and if
the repository is already downloaded the quick fix will do nothing.

Current implementation uses createDialogForPaths, which returns null in the
case that a package reuested is already installed. This function also
returns null in the case of an error and there is no way to know if the
returned null was caused by an error or by having a requested package
already present.

This CL adds a parameter (noOpMessage) to createDialog. When this parameter
is not null, showErrorDialog is called with this message if after resolving
dependencies there is nothing to install or uninstall.

Test: UnresolvedDependenciesReporterIntegrationTest and a new test in
SdkQuicfixUtilsTest to verify error dialog creation is called when a message
is passed to createDialogForPaths.

Patch 6: mofify createDialog insteadd of creating a similar function. This
reduces code duplication and also simplifies error handling.

Change-Id: Ia58f44fcf16055aeb97bb5624c0df17ab8cfdf4d