Replace usage of Dependencies with IdeLevel2Dependencies in adt/idea.
Replace all usages except for LintIdeProject.
Because LintIdeProject is sub class of Project class, and Project
class is defined in tools/base/sdk-common. Currently sdk-common
don't have access to Ide level2 model classes.

This is part of the effort to switch to DependencyGraphs in IDE.

Major difference is that IdeLevel2Dependencies returns flattened list
of level Library. So there is no need to traverse dependencies recursively.

Test: No new functionalities were added, existing tests should pass.
Change-Id: I991d4b233fe404e645623b94a52123bb21abf955