Move GutterIconCache#createIcon to GutterIconFactory
Currently, the logic for creating an Icon for display in the gutter
resides in private static methods in GutterIconCache and is hardcoded
to use fixed width and height limits on the resulting Icon. Having
this logic available outside the GutterIconCache class would avoid
code duplication for later gutter-related features
(see https://issuetracker.google.com/38346420) which need to generate
Icons for image resources according to various size constraints.

This CL fixes the issue by moving the logic for Icon creation to
a separate GutterIconFactory class, making the createIcon method
publicly accessible, and adding max height and width parameters to
the method signature.

Tests in GutterIconCacheTest.java related to this logic have been
moved to GutterIconFactoryTest.java.

Test: Covered by GutterIconFactoryTest.java
Change-Id: I70672cadcf9edeab098fae93e062832525f3ebc6