Add kotlin tests to TemplateTest
This CL tests the pre-converted kotlin templates by ensuring
a. That they build/compile w/o issues.
b. The generated directories do not contain any java files.
(This is true today since we expect the entire app to be in kotlin)

Note: This unearthed the issue with the kotlin plugin where
it doesn't like comma in the project-name. (Workaround included in the
test)
Test: CL adds tests to TemplateTest

Change-Id: Ifaf9d4023ca23031f3a7e2c39f02cee87da786d1