[AIA] Make provisioned devices cache more reliable
When a device is provisioned for instant apps, we put it in the cache as
a provisioned device and the cache is cleared each 24h. The problem is
that if the device is externally unprovisioned or if another device with
the same serialNumber (probably another emulator with the same
properties) is connected, it will be considered as provisioned and run
will fail.

This CL fixes this issue by verifying if the basic packages are
installed in the device and if they are not, the device will be
provisioned even if already in the cache.

Test: unit tests
Bug b/63144887

Change-Id: I4c974ee9b8ef558690ad1dc1a2ade05daeca2b48