Fix test-only modules support in Studio
Two main things were broken in support for test only modules:
1. The run configurations for test apps were being generated as unit
tests because the tests are in the main artifact of the variant and not
in the testArtifact.
2. The target module and variant apk was not being correctly provided.
It has to be queried in the PostBuildOutput model, since with config
splits we might not know the correct outputs at sync time.

This CL fixes these problems by changing some checks taking into account
test apps (1) and modifying the BuildAction used after build to also get
the ProjectBuildOutput of target modules.

Test: unit tests and reactivating test with test only module
Bug b/37137343

Change-Id: I502a374f1ff89c34f39fdabe0984ae097b326591